Move includes around to clean the source tree up a bit.
