+++ .syntastic_c_config
@@ -1,16 +1,3 @@
--I/home/pjones/devel/github.com/shim/master/Cryptlib/OpenSSL
--I/home/pjones/devel/github.com/shim/master/Cryptlib/OpenSSL/..
--I/home/pjones/devel/github.com/shim/master/Cryptlib/OpenSSL/../Include/
--I/home/pjones/devel/github.com/shim/master/Cryptlib/OpenSSL/crypto
--I/usr/lib/gcc/x86_64-redhat-linux/7/include
--I/home/pjones/devel/github.com/shim/master/Cryptlib/OpenSSL/../Include
--I/usr/include/efi
--I/usr/include/efi/x86_64
--I/usr/include/efi/protocol
--I/home/pjones/devel/github.com/shim/master/Cryptlib/OpenSSL/crypto/asn1
--I/home/pjones/devel/github.com/shim/master/Cryptlib/OpenSSL/crypto/evp
--I/home/pjones/devel/github.com/shim/master/Cryptlib/OpenSSL/crypto/modes
--I/home/pjones/devel/github.com/shim/master/Cryptlib/OpenSSL/crypto/include
 -DL_ENDIAN
 -D_CRT_SECURE_NO_DEPRECATE
 -D_CRT_NONSTDC_NO_DEPRECATE
@@ -29,18 +16,21 @@
 -Werror=sign-compare
 -ffreestanding
 -std=gnu89
--I/usr/lib/gcc/x86_64-redhat-linux/7/include
 -nostdinc
--I/home/pjones/devel/github.com/shim/master/Cryptlib
--I/home/pjones/devel/github.com/shim/master/Cryptlib/Include
--I/usr/include/efi
--I/usr/include/efi/x86_64
--I/usr/include/efi/protocol
--I/home/pjones/devel/github.com/shim/master/include
+-I/usr/lib/gcc/x86_64-redhat-linux/7/include
+-ICryptlib/
+-ICryptlib/Include/
+-ICryptlib/OpenSSL/
+-ICryptlib/OpenSSL/crypto/
+-I/usr/include/efi/
+-I/usr/include/efi/x86_64/
+-I/usr/include/efi/protocol/
+-ICryptlib/OpenSSL/crypto/asn1/
+-ICryptlib/OpenSSL/crypto/evp/
+-ICryptlib/OpenSSL/crypto/modes/
+-ICryptlib/OpenSSL/crypto/include/
 -iquote
-/home/pjones/devel/github.com/shim/master
--iquote
-/home/pjones/devel/github.com/shim/master
+.
 -mno-mmx
 -mno-sse
 -mno-red-zone
