+++ netboot.c
@@ -34,9 +34,8 @@
  */
 
 #include "shim.h"
+
 #include <string.h>
-#include "netboot.h"
-#include "str.h"
 
 #define ntohs(x) __builtin_bswap16(x)	/* supported both by GCC and clang */
 #define htons(x) ntohs(x)
